// openapi/requests/types.gen.ts
export type Id = number;

export type UrlType = string;
// ...
export type Rating = number;

export type ReviewContent = string;

export type CreateReviewRequestBody = {
  productId: Id;
  /**
   * 이미지 URL 입니다. 최대 3개까지 가능합니다.
   */
  images?: Array<UrlType>;
  content: ReviewContent;
  rating: Rating;
};

export type UpdateReviewRequestBody = {
  /**
   * 기존 이미지를 유지하려면 id를, 새로운 이미지를 추가하려면 source를 넣어주세요. <br />
   * 요청에 포함되지 않는 기존 이미지는 삭제됩니다.
   */
  images?: Array<
    | {
        id: Id;
      }
    | {
        source: UrlType;
      }
  >;
  content?: ReviewContent;
  rating?: Rating;
};
// ...
export type CreateReviewData = {
  body: CreateReviewRequestBody;
  path: {
    teamId: string;
  };
};
// ...
export type UpdateReviewData = {
  body: UpdateReviewRequestBody;
  path: {
    reviewId: Id;
    teamId: string;
  };
};